+++ src/main/java/com/solvd/laba/bank/models/Debit_in_card.java
@@ -1,9 +1,9 @@
 package com.solvd.laba.bank.models;
 
 public class Debit_in_card {
-	private Long card_id;
-	private Long debit_id;
-	public Debit_in_card(Long card_id, Long debit_id) {
+	private Integer card_id;
+	private Integer debit_id;
+	public Debit_in_card(Integer card_id, Integer debit_id) {
 		super();
 		this.card_id = card_id;
 		this.debit_id = debit_id;
@@ -11,16 +11,16 @@ public class Debit_in_card {
 	public Debit_in_card() {
 		super();
 	}
-	public Long getCard_id() {
+	public Integer getCard_id() {
 		return card_id;
 	}
-	public void setCard_id(Long card_id) {
+	public void setCard_id(Integer card_id) {
 		this.card_id = card_id;
 	}
-	public Long getDebit_id() {
+	public Integer getDebit_id() {
 		return debit_id;
 	}
-	public void setDebit_id(Long debit_id) {
+	public void setDebit_id(Integer debit_id) {
 		this.debit_id = debit_id;
 	}
